feat: implement project membership policies and enhance authorization checks in controllers

+++ backend_app/app/Http/Controllers/Project/ProjectMember/ProjectMemberController.php
@@ -11,6 +11,8 @@ use App\Http\Resources\Project\ProjectMember\ProjectMemberResource;
 use App\Models\Project;
 use App\Models\ProjectMember;
 use App\Models\Workspace;
+use DB;
+use Gate;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
@@ -23,6 +25,8 @@ class ProjectMemberController extends Controller
      */
     public function index(Workspace $workspace, Project $project): JsonResource
     {
+        Gate::authorize('viewAny', [ProjectMember::class, $project]);
+
         return new ProjectMemberCollection(
             $project->memberships()
                 ->with(['project', 'user'])
@@ -33,14 +37,21 @@ class ProjectMemberController extends Controller
 
     /**
      * Store a newly created resource in storage.
+     * @throws Throwable
      */
     public function store(StoreProjectMember $request, Workspace $workspace, Project $project): JsonResource
     {
-        $member = $project->memberships()->make($request->validated());
+        Gate::authorize('create', [ProjectMember::class, $project]);
 
-        $member->forceFill([
-            'approval_rank' => $member->project_role->approvalRank(),
-        ])->save();
+        $member = DB::transaction(function () use ($request, $workspace, $project) {
+            $member = $project->memberships()->make($request->validated());
+
+            $member->forceFill([
+                'approval_rank' => $member->project_role->approvalRank(),
+            ])->saveOrFail();
+
+            return $member;
+        });
 
         return new ProjectMemberResource($member->load(['project', 'user']));
     }
@@ -50,6 +61,8 @@ class ProjectMemberController extends Controller
      */
     public function show(Workspace $workspace, Project $project, ProjectMember $membership): JsonResource
     {
+        Gate::authorize('view', $membership);
+
         return new ProjectMemberResource($membership->load(['project', 'user']));
     }
 
@@ -63,6 +76,8 @@ class ProjectMemberController extends Controller
         Project             $project,
         ProjectMember       $membership): JsonResource
     {
+        Gate::authorize('update', $membership);
+
         $membership->updateOrFail($request->validated());
 
         return new ProjectMemberResource($membership->load(['project', 'user']));
@@ -74,6 +89,8 @@ class ProjectMemberController extends Controller
      */
     public function destroy(Workspace $workspace, Project $project, ProjectMember $membership): JsonResponse
     {
+        Gate::authorize('delete', $membership);
+
         $membership->deleteOrFail();
 
         return response()->json(status: 204);
